+++ Engine/scenes/level/lvl_block.h
@@ -1,6 +1,6 @@
 /*
  * Platformer Game Engine by Wohlstand, a free platform for game making
- * Copyright (c) 2014 Vitaly Novichkov <admin@wohlnet.ru>
+ * Copyright (c) 2015 Vitaly Novichkov <admin@wohlnet.ru>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -81,7 +81,7 @@ public:
 
     //float posX();
     //float posY();
-    void render(float camX, float camY);
+    void render(double camX, double camY);
 private:
     void drawPiece(QRectF target, QRectF block, QRectF texture);
 };
